requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
MarkupSafe==0.21
SQLAlchemy==0.9.4
Werkzeug==0.9.4
gevent==1.0
greenlet==0.4.2
itsdangerous==0.24
psycopg2==2.5.2
pycrypto==2.6.1
pysqlite==2.6.3
redis==2.9.1
wsgiref==0.1.2
